+++ src/main/java/io/kristaxlab/notion/model/common/NotionObject.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import io.kristaxlab.notion.model.BaseNotionObject;
 import io.kristaxlab.notion.model.block.Block;
+import io.kristaxlab.notion.model.comment.Comment;
 import io.kristaxlab.notion.model.database.Database;
 import io.kristaxlab.notion.model.datasource.DataSource;
 import io.kristaxlab.notion.model.file.FileUpload;
@@ -22,7 +23,8 @@ import lombok.Setter;
   @JsonSubTypes.Type(value = DataSource.class, name = "data_source"),
   @JsonSubTypes.Type(value = Page.class, name = "page"),
   @JsonSubTypes.Type(value = FileUpload.class, name = "file_upload"),
-  @JsonSubTypes.Type(value = Block.class, name = "block")
+  @JsonSubTypes.Type(value = Block.class, name = "block"),
+  @JsonSubTypes.Type(value = Comment.class, name = "comment")
 })
 @Getter
 @Setter