+++ src/testIntegration/java/suites/QuickCheck.java
@@ -3,18 +3,18 @@ package suites;
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.Suite;
 import org.junit.platform.suite.api.SuiteDisplayName;
+import tests.blocks.IT11_Blocks_ParagraphCRUD;
 import tests.blocks.IT13_Blocks_SeveralBlocksAtOnce;
 import tests.blocks.IT16_Blocks_InsertToPosition;
-import tests.blocks.IT22_Blocks_ParagraphCRUD;
 import tests.fileuploads.IT3_FileUploads_SinglePartFile;
 import tests.fileuploads.IT5_FileUploads_ListAll;
 import tests.pages.*;
-import tests.users.IT32_Users_ListAll;
+import tests.users.IT2_Users_ListAll;
 
 @Suite
 @SuiteDisplayName("QuickCheck")
 @SelectClasses({
-  IT22_Blocks_ParagraphCRUD.class,
+  IT11_Blocks_ParagraphCRUD.class,
   IT13_Blocks_SeveralBlocksAtOnce.class,
   IT16_Blocks_InsertToPosition.class,
   IT1_Pages_CRUD.class,
@@ -25,6 +25,6 @@ import tests.users.IT32_Users_ListAll;
   IT10_Pages_RichTextPropertyPaginated.class,
   IT3_FileUploads_SinglePartFile.class,
   IT5_FileUploads_ListAll.class,
-  IT32_Users_ListAll.class
+  IT2_Users_ListAll.class
 })
 public class QuickCheck {}